+++ plugins/qeta/dev/ComponentPage.tsx
@@ -1,0 +1,20 @@
+import { Content, Page } from '@backstage/core-components';
+import { Container } from '@material-ui/core';
+import { QuestionsContainer } from '../src';
+import React from 'react';
+
+export const ComponentPage = () => {
+  return (
+    <Page themeId="home">
+      <Content>
+        <Container>
+          <QuestionsContainer
+            entity="component:default/artist-web"
+            showTitle
+            showAskButton
+          />
+        </Container>
+      </Content>
+    </Page>
+  );
+};