+++ plugins/qeta/src/utils/utils.ts
@@ -1,11 +1,5 @@
 import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 
-export const getEntityUrl = (entity: Entity) => {
-  return `/catalog/${entity.metadata.namespace ?? 'default'}/${entity.kind}/${
-    entity.metadata.name
-  }`.toLowerCase();
-};
-
 export const formatEntityName = (username: string) => {
   const plainName = username.split(/[/:]+/).pop();
   return plainName